Commited with updations

+++ src/main/java/com/kgisl/sb1/entity/Person.java
@@ -1,8 +1,6 @@
 package com.kgisl.sb1.entity;
 
-// import org.springframework.data.annotation.Id;
-
 import org.springframework.data.annotation.Id;
 
-public record Person(@Id Long id, String uname, String email) {
+public record Person(@Id int id, String uname, String email) {
 }
